First person camera that rotates in y rotation
-Implemented new rotate function
-Implemented new move function
-Implemented a hover over function that moves mouse to center and rotates camera
-Move is now bound to wasd

+++ PizzaBoi/Quaternion.h
@@ -10,10 +10,10 @@ public:
 	Quaternion conjugate();
 	Quaternion mult(Quaternion B);
 
-	double x;
-	double y;
-	double z;
-	double w;
+	GLfloat x;
+	GLfloat y;
+	GLfloat z;
+	GLfloat w;
 
 	
 };
